rename browser connection error hints file
