+++ src/browser/connection/error-hints.ts
